// FrontEnd/src/interfaces/ITVShowDetails.ts
export default interface ITVShowDetails {
    "countries": string[];
    "description": string;
    "creators": string[];
    "genres": string[];
    "imdb_id": string;
    "imdb_rating": string;
    "language": string[];
    "networks": string[];
    "popularity": string;
    "production_companies": string[];
    "rated": string;
    "release_date": string;
    "runtime": number;
    "stars": string[];
    "status": string;
    "status_message": string;
    "title": string;
    "vote_count": string;
    "year_started": number;
    "youtube_trailer_key": string;
    "fanart": string;
    "poster": string;
}